requirements.txt
Flask==3.0.0
WeasyPrint==60.1
PyYAML==6.0.1
Jinja2==3.1.2